Tests on box2d of VOC2012Detection.

| cases               | before           | after             |
| ---------------     | ---------------- | ----------------  |
| box2d.internal_type | 59.5 µs ± 104 ns | 54.1 µs ± 99.7 ns |
| box2d.to_builtin()  | 145 µs ± 14.5 µs | 124 µs ± 292 ns   |
| box2d._data         | 90.2 µs ± 113 ns | 80.5 µs ± 3.59 µs |