@GravitiRob 2022-06-24
Task linked: [CU-2j4fdy5 Portex Factory 类型中使用字典代替 **kwargs 进行传参](https://app.clickup.com/t/2j4fdy5)